docker_airflow/requirements.txt
boto3==1.34.49
botocore==1.34.49
bs4==0.0.2
config==0.5.1
contourpy==1.2.0
google-ai-generativelanguage==0.6.2
google-api-core==2.18.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
openai==1.23.6
PyMySQL==1.1.0
pymysql-pooling==1.0.6
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
schedule==1.2.1
scipy==1.12.0
selenium==4.19.0
tiktoken==0.6.0
webdriver-manager==4.0.1
Werkzeug==3.0.2
widgetsnbextension==4.0.10
